+++ backend/typescript/seeders/2021.11.25T00.54.43.sample-series.ts
@@ -1,0 +1,20 @@
+import { Seeder } from "../umzug";
+
+const seedSeries = [
+  {
+    id: 1,
+    name: "Bloom Trilogy",
+    createdAt: new Date(Date.now()),
+    updatedAt: new Date(Date.now()),
+  },
+];
+
+export const up: Seeder = async ({ context: sequelize }) => {
+  await sequelize.getQueryInterface().bulkInsert("series", seedSeries);
+};
+
+export const down: Seeder = async ({ context: sequelize }) => {
+  await sequelize
+    .getQueryInterface()
+    .bulkDelete("series", { id: seedSeries.map((u) => u.id) });
+};
