+++ backend/typescript/seeders/2021.11.25T00.55.24.sample-publishers.ts
@@ -1,0 +1,21 @@
+import { Seeder } from "../umzug";
+
+const seedPublishers = [
+  {
+    id: 1,
+    full_name: "HarperCollins Canada",
+    publish_year: 2020,
+    createdAt: new Date(Date.now()),
+    updatedAt: new Date(Date.now()),
+  },
+];
+
+export const up: Seeder = async ({ context: sequelize }) => {
+  await sequelize.getQueryInterface().bulkInsert("publishers", seedPublishers);
+};
+
+export const down: Seeder = async ({ context: sequelize }) => {
+  await sequelize
+    .getQueryInterface()
+    .bulkDelete("publishers", { id: seedPublishers.map((u) => u.id) });
+};
